+++ CHANGELOG.md
@@ -2,5 +2,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+### Changed
+* Improve exception hierarchy.
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+
 ## [0.1.0] - 2021-07-26
 Initial version.